{"types":"C","runs":13}

+++ problem12.java
@@ -25,7 +25,7 @@ public class problem12 extends GraphicsProgram{
 	
 	private void drawWall() {
 		GRect wall = new GRect(HOUSE_WIDTH, HOUSE_HEIGHT);
-		int x = (int) (getWidth() / 2 - HOUSE_WIDTH / 2);
+		double x = (int) (getWidth() / 2 - HOUSE_WIDTH / 2);
 		double y = getHeight() - HOUSE_HEIGHT;
 		add(wall, x, y);
 	}
